combine two go commands
